[FIX] Split files analytic.py
