+++ analytic_tag_dimension_enhanced/__init__.py
@@ -1,4 +1,3 @@
-# Copyright 2019 Ecosoft Co., Ltd (http://ecosoft.co.th/)
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html)
 
 from . import models
