+++ analytic_tag_dimension_enhanced/models/__init__.py
@@ -1,5 +1,5 @@
-# Copyright 2019 Ecosoft Co., Ltd (http://ecosoft.co.th/)
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html)
 
-from . import analytic
 from . import analytic_dimension_line
+from . import account_analytic_dimension
+from . import account_analytic_tag
